refactor: merged CategoriesTree and NavigationTree into SideBarTree to keep thing DRY
